Refactored upgrade-nudge to only receive essential parameters
